Rename gem from verify_nexmo_signature to nexmo_rack
