+++ CHANGES.md
@@ -1,3 +1,7 @@
+# 0.2.0
+
+* Renamed from `verify_nexmo_signature` to `nexmo_rack`
+
 # 0.1.0
 
 * First version! 🎉
