+++ spec/nexmo_rack/verify_signature_spec.rb
@@ -1,20 +1,20 @@
 # frozen_string_literal: true
 
-require_relative '../lib/verify_nexmo_signature'
+require_relative '../../lib/nexmo_rack'
 
-describe VerifyNexmoSignature do
+describe Nexmo::Rack::VerifySignature do
   before do
     ENV['NEXMO_API_SIGNATURE'] = 'secret'
   end
 
   let(:app) do
     Rack::Builder.new do
-      use VerifyNexmoSignature::Middleware
+      use Nexmo::Rack::VerifySignature
       run(->(env) { [200, env, "app"] })
     end 
-  end  
+  end
 
-  let(:app_valid_params) do
+  let(:base_params) do
     {
       'message-timestamp' => '2013-11-21 17:31:42',
       'messageId' => '030000002A264B8B',
@@ -23,30 +23,19 @@ describe VerifyNexmoSignature do
       'timestamp' => '1461605396',
       'to' => '14849970568',
       'type' => 'text',
-      'sig' => '56859cb8e7ba1da0bb9c8ec0e703feb0'
     }
+  end
+
+  let(:app_valid_params) do
+    base_params.merge({
+      'sig' => '56859cb8e7ba1da0bb9c8ec0e703feb0'
+    })
   end
 
   let(:app_invalid_params) do
-    {
-      'message-timestamp' => '2013-11-21 17:31:42',
-      'messageId' => '030000002A264B8B',
-      'msisdn' => '14843472194',
-      'text' => 'Test',
-      'timestamp' => '1461605396',
-      'to' => '14849970568',
-      'type' => 'text',
+    base_params.merge({
       'sig' => 'xxxx'
-    }
-  end
-
-  let :middleware do
-    VerifyNexmoSignature::Middleware.new(app)
-  end
-
-  it 'returns a 200 HTTP response code if it is called directly and not a POST request' do
-    response = middleware.call(app_invalid_params)
-    expect(response[0]).to eq(200)
+    })
   end
 
   it 'returns a 200 HTTP response for a POST request with a valid verification' do
